add todoReducer
